ShareServiceConnectionCommandHandler
